added helpers for day to number converter vice versa, fixed restaurant bug where only 1 restaurant would show up when user trying to get restaurant open at specific day, and last fixed getexistseat by adding status=Accepted as a where criteria
